move to antitypical result
