+++ Cartfile
@@ -1,3 +1,4 @@
 github "Alamofire/Alamofire" ~> 3.0
 github "ReactiveCocoa/ReactiveCocoa" "v4.0.0-alpha.4"
 github "ReactiveX/RxSwift" "2.0.0-beta.4"
+github "antitypical/Result" ~> 1.0
