+++ Moya.podspec
@@ -23,6 +23,7 @@ Pod::Spec.new do |s|
   s.subspec "Core" do |ss|
     ss.source_files  = "Source/*.swift", "Source/Plugins/*swift"
     ss.dependency "Alamofire", "~> 3.0"
+    ss.dependency "Result", "~> 1.0"
     ss.framework  = "Foundation"
   end
