+++ Source/Moya.swift
@@ -1,33 +1,12 @@
 import Foundation
-import Alamofire
+import Result
 
 /// Closure to be executed when a request has completed.
-public typealias Completion = (result: Moya.Result<Moya.Response, Moya.Error>) -> ()
+public typealias Completion = (result: Result<Moya.Response, Moya.Error>) -> ()
 
 /// Represents an HTTP method.
 public enum Method: String {
     case GET, POST, PUT, DELETE, OPTIONS, HEAD, PATCH, TRACE, CONNECT
-}
-
-/// Choice of parameter encoding.
-public enum ParameterEncoding {
-    case URL
-    case JSON
-    case PropertyList(NSPropertyListFormat, NSPropertyListWriteOptions)
-    case Custom((URLRequestConvertible, [String: AnyObject]?) -> (NSMutableURLRequest, NSError?))
-    
-    internal var toAlamofire: Alamofire.ParameterEncoding {
-        switch self {
-        case .URL:
-            return .URL
-        case .JSON:
-            return .JSON
-        case .PropertyList(let format, let options):
-            return .PropertyList(format, options)
-        case .Custom(let closure):
-            return .Custom(closure)
-        }
-    }
 }
 
 public enum StubBehavior {
@@ -75,7 +54,7 @@ public class MoyaProvider<Target: TargetType> {
     public init(endpointClosure: EndpointClosure = MoyaProvider.DefaultEndpointMapping,
         requestClosure: RequestClosure = MoyaProvider.DefaultRequestMapping,
         stubClosure: StubClosure = MoyaProvider.NeverStub,
-        manager: Manager = Alamofire.Manager.sharedInstance,
+        manager: Manager = Manager.sharedInstance,
         plugins: [PluginType] = []) {
             
             self.endpointClosure = endpointClosure
@@ -197,20 +176,20 @@ internal extension MoyaProvider {
         return {
             if (token.canceled) {
                 let error = Moya.Error.Underlying(NSError(domain: NSURLErrorDomain, code: NSURLErrorCancelled, userInfo: nil))
-                plugins.forEach { $0.didReceiveResponse(Moya.Result(failure: error), target: target) }
-                completion(result: Result(failure: error))
+                plugins.forEach { $0.didReceiveResponse(.Failure(error), target: target) }
+                completion(result: .Failure(error))
                 return
             }
             
             switch endpoint.sampleResponseClosure() {
             case .NetworkResponse(let statusCode, let data):
                 let response = Moya.Response(statusCode: statusCode, data: data, response: nil)
-                plugins.forEach { $0.didReceiveResponse(Moya.Result(success: response), target: target) }
-                completion(result: Moya.Result(success: response))
+                plugins.forEach { $0.didReceiveResponse(.Success(response), target: target) }
+                completion(result: .Success(response))
             case .NetworkError(let error):
                 let error = Moya.Error.Underlying(error)
-                plugins.forEach { $0.didReceiveResponse(Moya.Result(failure: error), target: target) }
-                completion(result: Moya.Result(failure: error))
+                plugins.forEach { $0.didReceiveResponse(.Failure(error), target: target) }
+                completion(result: .Failure(error))
             }
         }
     }
@@ -222,55 +201,19 @@ internal extension MoyaProvider {
     }
 }
 
-private func convertResponseToResult(response: NSHTTPURLResponse?, data: NSData?, error: NSError?) -> Moya.Result<Moya.Response, Moya.Error> {
+private func convertResponseToResult(response: NSHTTPURLResponse?, data: NSData?, error: NSError?) ->
+    Result<Moya.Response, Moya.Error> {
     switch (response, data, error) {
     case let (.Some(response), .Some(data), .None):
         let response = Moya.Response(statusCode: response.statusCode, data: data, response: response)
-        return Moya.Result(success: response)
+        return .Success(response)
     case let (.None, .None, .Some(error)):
         let error = Moya.Error.Underlying(error)
-        return Moya.Result(failure: error)
+        return .Failure(error)
     default:
         let error = Moya.Error.Underlying(NSError(domain: NSURLErrorDomain, code: NSURLErrorUnknown, userInfo: nil))
-        return Moya.Result(failure: error)
+        return .Failure(error)
     }
-}
-
-/// Internal token that can be used to cancel requests
-internal final class CancellableToken: Cancellable , CustomDebugStringConvertible{
-    let cancelAction: () -> Void
-    let request : Request?
-    private(set) var canceled: Bool = false
-    
-    private var lock: OSSpinLock = OS_SPINLOCK_INIT
-    
-    func cancel() {
-        OSSpinLockLock(&lock)
-        defer { OSSpinLockUnlock(&lock) }
-        guard !canceled else { return }
-        canceled = true
-        cancelAction()
-    }
-    
-    init(action: () -> Void){
-        self.cancelAction = action
-        self.request = nil
-    }
-    
-    init(request : Request){
-        self.request = request
-        self.cancelAction = {
-            request.cancel()
-        }
-    }
-    
-    var debugDescription: String {
-        guard let request = self.request else {
-            return "Empty Request"
-        }
-        return request.debugDescription
-    }
-    
 }
 
 private struct CancellableWrapper: Cancellable {
@@ -282,6 +225,3 @@ private struct CancellableWrapper: Cancellable {
         innerCancellable?.cancel()
     }
 }
-
-/// Make the Alamofire Request type conform to our type, to prevent leaking Alamofire to plugins.
-extension Request: RequestType { }
