+++ Source/Plugins/NetworkActivityPlugin.swift
@@ -1,4 +1,5 @@
 import Foundation
+import Result
 
 /// Network activity change notification type.
 public enum NetworkActivityChangeType {
