+++ Source/ReactiveCocoa/SignalProducer+Moya.swift
@@ -1,5 +1,6 @@
 import Foundation
 import ReactiveCocoa
+import Result
 
 /// Extension for processing raw NSData generated by network access.
 extension SignalProducerType where Value == Response, Error == Moya.Error {
@@ -7,55 +8,55 @@ extension SignalProducerType where Value == Response, Error == Moya.Error {
     /// Filters out responses that don't fall within the given range, generating errors when others are encountered.
     public func filterStatusCodes(range: ClosedInterval<Int>) -> SignalProducer<Value, Error> {
         return producer.flatMap(.Latest) { response -> SignalProducer<Value, Error> in
-            return unwrapThrowable { try response.filterStatusCodes(range) }
+            return SignalProducer(result: Result { try response.filterStatusCodes(range) })
         }
     }
     
     public func filterStatusCode(code: Int) -> SignalProducer<Value, Error> {
         return producer.flatMap(.Latest) { response -> SignalProducer<Value, Error> in
-            return unwrapThrowable { try response.filterStatusCode(code) }
+            return SignalProducer(result: Result { try response.filterStatusCode(code) })
         }
     }
     
     public func filterSuccessfulStatusCodes() -> SignalProducer<Value, Error> {
         return producer.flatMap(.Latest) { response -> SignalProducer<Value, Error> in
-            return unwrapThrowable { try response.filterSuccessfulStatusCodes() }
+            return SignalProducer(result: Result { try response.filterSuccessfulStatusCodes() })
         }
     }
     
     public func filterSuccessfulStatusAndRedirectCodes() -> SignalProducer<Value, Error> {
         return producer.flatMap(.Latest) { response -> SignalProducer<Value, Error> in
-            return unwrapThrowable { try response.filterSuccessfulStatusAndRedirectCodes() }
+            return SignalProducer(result: Result { try response.filterSuccessfulStatusAndRedirectCodes() })
         }
     }
     
     /// Maps data received from the signal into an Image. If the conversion fails, the signal errors.
     public func mapImage() -> SignalProducer<Image, Error> {
         return producer.flatMap(.Latest) { response -> SignalProducer<Image, Error> in
-            return unwrapThrowable { try response.mapImage() }
+            return SignalProducer(result: Result { try response.mapImage() })
         }
     }
     
     /// Maps data received from the signal into a JSON object. If the conversion fails, the signal errors.
     public func mapJSON() -> SignalProducer<AnyObject, Error> {
         return producer.flatMap(.Latest) { response -> SignalProducer<AnyObject, Error> in
-            return unwrapThrowable { try response.mapJSON() }
+            return SignalProducer(result: Result { try response.mapJSON() })
         }
     }
     
     /// Maps data received from the signal into a String. If the conversion fails, the signal errors.
     public func mapString() -> SignalProducer<String, Error> {
         return producer.flatMap(.Latest) { response -> SignalProducer<String, Error> in
-            return unwrapThrowable { try response.mapString() }
+            return SignalProducer(result: Result { try response.mapString() })
         }
     }
 }
-
-/// Maps throwable to SignalProducer
-private func unwrapThrowable<T>(throwable: () throws -> T) -> SignalProducer<T, Error> {
-    do {
-        return SignalProducer(value: try throwable())
-    } catch {
-        return SignalProducer(error: error as! Moya.Error)
-    }
-}
+//
+///// Maps throwable to SignalProducer
+//private func unwrapThrowable<T>(throwable: () throws -> T) -> SignalProducer<T, Error> {
+//    do {
+//        return SignalProducer(value: try throwable())
+//    } catch {
+//        return SignalProducer(error: error as! Moya.Error)
+//    }
+//}
